requirements.txt
numpy~=2.3.0
scipy~=1.16.0
seaborn~=0.13.2
matplotlib~=3.10
pandas~=2.3.1
h5py~=3.14.0
tqdm~=4.67
pyyaml~=6.0.2
